Consolidate FillAndKill/FillOrKill into ImmediateOrCancel
CME's own FIX protocol doesn't have a distinct FOK time-in-force: it's
IOC (tag 59=3) with MinQty (tag 110) set equal to OrderQty - FOK and FAK
are the same wire-level concept. Verified the consolidation is
mechanically sound here too, not just analogous: HasSufficientLiquidity
already walks the exact same liquidity Match() would consume, so a gate
that passes with MinQuantity == the order's full Quantity guarantees the
whole order fills - exactly FOK's behavior, through the same code path
as FillAndKill+MinQuantity, with no special-casing.

This removes real duplication (two liquidity-gate checks in both
CreateOrder and TriggerStops that are now provably the same check) and
one dead-code path: OrderCancelledReason.FillOrKillNotFilled becomes
unreachable once consolidated (a FOK-equivalent order that passes its
gate always ends up fully Filled, never partially-filled-then-cancelled),
so it's deleted outright along with OrderRejectedReason
.InsufficientLiquidityForFillOrKill, rather than kept as dead code.

Also renamed QuantityOutOfRange to MinQuantityOutOfRange for clarity,
and consolidated the three now-overlapping test files (FillAndKillTests,
FillOrKillTests, MinQuantityTests) into one ImmediateOrCancelTests file.
