+++ Circus/OrderBook/OrderRejectedReason.cs
@@ -17,7 +17,6 @@ namespace Circus.OrderBook
         TriggerPriceMustBeLessThanPrice,
         TriggerPriceMustBeGreaterThanPrice,
         NoChange,
-        InsufficientLiquidityForFillOrKill,
         InvalidExpireDate,
         ClientOrderIdRequired,
         ClientOrderIdTooLong,
@@ -25,7 +24,7 @@ namespace Circus.OrderBook
         CompanyIdTooLong,
         SelfMatchPreventionIdTooLong,
         PriceOutsideBands,
-        QuantityOutOfRange,
+        MinQuantityOutOfRange,
         InsufficientLiquidityForMinQuantity,
         // PendingCancelOrReplace,
         // PriceExceedsCurrentPrice,
