+++ Circus/OrderValidity.cs
@@ -12,16 +12,18 @@ namespace Circus
             public required DateOnly Date { get; init; }
         }
 
-        // MinQuantity is meaningful only for FillAndKill (CME's MinQty is documented as an
-        // IOC-specific qualifier: if at least this much can fill immediately, the order proceeds
-        // as an ordinary FillAndKill; otherwise nothing fills at all, not even a partial below the
-        // minimum). Carrying it here rather than as a field elsewhere makes that pairing structural
-        // - there's no way to construct a MinQuantity that isn't attached to a FillAndKill.
-        public sealed record FillAndKill : OrderValidity
+        // Consolidates what CME's own FIX protocol treats as one concept: IOC (tag 59=3) with an
+        // optional MinQty (tag 110) qualifier - there's no separate "FOK" time-in-force at the
+        // wire level. MinQuantity null (or less than the order's own Quantity) behaves like
+        // classic IOC/FillAndKill: fills whatever's immediately available, cancels the remainder,
+        // no minimum required. Setting MinQuantity equal to the order's Quantity reproduces
+        // FillOrKill exactly - HasSufficientLiquidity walks the same liquidity Match() would
+        // actually consume, so if the gate passes at MinQuantity == Quantity the whole order is
+        // guaranteed to fill; if it fails, the order is rejected before ever entering the book.
+        // No special-casing needed for either.
+        public sealed record ImmediateOrCancel : OrderValidity
         {
             public int? MinQuantity { get; init; }
         }
-
-        public sealed record FillOrKill : OrderValidity;
     }
 }
